+++ internal/provider/model_catalystcenter_fabric_multicast_virtual_networks.go
@@ -68,9 +68,6 @@ func (data FabricMulticastVirtualNetworks) getPath() string {
 // End of section. //template:end getPath
 
 // Section below is generated&owned by "gen/generator.go". //template:begin getFallbackPath
-func (data FabricMulticastVirtualNetworks) getFallbackPath() string {
-	return ""
-}
 
 // End of section. //template:end getFallbackPath