+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+## 0.4.4 (unreleased)
+
+- Add `catalystcenter_integrate_ise` resource to accept Cisco ISE server certificate for Cisco ISE server integration
+
 ## 0.4.3
 
 - Add `catalystcenter_fabric_devices` resource and data_source to manage multiple fabric devices under same fabric site within single resource