+++ docs/guides/changelog.md
@@ -7,6 +7,10 @@ description: |-
 
 # Changelog
 
+## 0.4.4 (unreleased)
+
+- Add `catalystcenter_integrate_ise` resource to accept Cisco ISE server certificate for Cisco ISE server integration
+
 ## 0.4.3
 
 - Add `catalystcenter_fabric_devices` resource and data_source to manage multiple fabric devices under same fabric site within single resource